Collapse runtime client and facade
