+++ src/vercel/unstable/sandbox/__init__.py
@@ -2,7 +2,19 @@
 
 from collections.abc import AsyncIterator, Mapping
 
-from vercel._internal.unstable.context import get_active_session
+from vercel._internal.unstable.sandbox.async_runtime import (
+    CreateSandboxOperation,
+    Sandbox,
+    SandboxCommand,
+    SandboxRuntimeSession,
+    Snapshot,
+    create_sandbox_operation as _create_sandbox_operation,
+    get_sandbox as _get_sandbox,
+    get_snapshot as _get_snapshot,
+    query_sandboxes as _query_sandboxes,
+    query_sessions as _query_sessions,
+    query_snapshots as _query_snapshots,
+)
 from vercel._internal.unstable.sandbox.errors import (
     SandboxApiError,
     SandboxCleanupError,
@@ -12,12 +24,6 @@ from vercel._internal.unstable.sandbox.errors import (
     SandboxResponseError,
     SandboxStreamError,
     SandboxTerminalStateError,
-)
-from vercel._internal.unstable.sandbox.handles import (
-    Sandbox,
-    SandboxCommand,
-    SandboxRuntimeSession,
-    Snapshot,
 )
 from vercel._internal.unstable.sandbox.models import (
     DurationInput,
@@ -39,11 +45,8 @@ from vercel._internal.unstable.sandbox.models import (
     TarballSource,
     WriteFile,
 )
-from vercel._internal.unstable.sandbox.operations import (
-    CreateSandboxOperation,
-    create_sandbox_operation,
-)
 from vercel._internal.unstable.sandbox.options import SandboxServiceOptions
+from vercel._internal.unstable.session import get_active_session
 
 from . import sync
 
@@ -64,7 +67,8 @@ def create_sandbox(
     snapshot_expiration: DurationInput = None,
     snapshot_retention: SnapshotRetention | None = None,
 ) -> CreateSandboxOperation:
-    return create_sandbox_operation(
+    return _create_sandbox_operation(
+        get_active_session().sandbox_service(),
         project_id=project_id,
         name=name,
         runtime=runtime,
@@ -88,15 +92,12 @@ async def get_sandbox(
     resume: bool = True,
     include_system_routes: bool | None = None,
 ) -> Sandbox:
-    return (
-        await get_active_session()
-        .sandbox_service()
-        .get_sandbox(
-            name=name,
-            project_id=project_id,
-            resume=resume,
-            include_system_routes=include_system_routes,
-        )
+    return await _get_sandbox(
+        get_active_session().sandbox_service(),
+        name=name,
+        project_id=project_id,
+        resume=resume,
+        include_system_routes=include_system_routes,
     )
 
 
@@ -107,15 +108,12 @@ def query_sandboxes(
     page_size: int | None = None,
     cursor: str | None = None,
 ) -> AsyncIterator[Sandbox]:
-    return (
-        get_active_session()
-        .sandbox_service()
-        .query_sandboxes(
-            query=query,
-            project_id=project_id,
-            page_size=page_size,
-            cursor=cursor,
-        )
+    return _query_sandboxes(
+        get_active_session().sandbox_service(),
+        query=query,
+        project_id=project_id,
+        page_size=page_size,
+        cursor=cursor,
     )
 
 
@@ -127,16 +125,13 @@ def query_sessions(
     cursor: str | None = None,
     sort_order: str | None = None,
 ) -> AsyncIterator[SandboxRuntimeSession]:
-    return (
-        get_active_session()
-        .sandbox_service()
-        .query_sessions(
-            project_id=project_id,
-            name=name,
-            page_size=page_size,
-            cursor=cursor,
-            sort_order=sort_order,
-        )
+    return _query_sessions(
+        get_active_session().sandbox_service(),
+        project_id=project_id,
+        name=name,
+        page_size=page_size,
+        cursor=cursor,
+        sort_order=sort_order,
     )
 
 
@@ -148,21 +143,18 @@ def query_snapshots(
     cursor: str | None = None,
     sort_order: str | None = None,
 ) -> AsyncIterator[Snapshot]:
-    return (
-        get_active_session()
-        .sandbox_service()
-        .query_snapshots(
-            project_id=project_id,
-            name=name,
-            page_size=page_size,
-            cursor=cursor,
-            sort_order=sort_order,
-        )
+    return _query_snapshots(
+        get_active_session().sandbox_service(),
+        project_id=project_id,
+        name=name,
+        page_size=page_size,
+        cursor=cursor,
+        sort_order=sort_order,
     )
 
 
 async def get_snapshot(*, snapshot_id: str) -> Snapshot:
-    return await get_active_session().sandbox_service().get_snapshot(snapshot_id=snapshot_id)
+    return await _get_snapshot(get_active_session().sandbox_service(), snapshot_id=snapshot_id)
 
 
 __all__ = [
